pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "peewee>=3.17",
    "psycopg2-binary>=2.9",
    "psutil>=6.0",
    "python-dotenv>=1.0",
    "redis>=5.0",
    "prometheus-client>=0.24.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=50"
testpaths = ["tests"]
